+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - note subfolders deeper than 2nd level can now be used
   (for [Issue #246](https://github.com/pbek/QOwnNotes/issues/246))
 - note subfolders can now be renamed in the note subfolder tree
+- now the root note folder cannot be removed in the note subfolder tree  
+  (for [Issue #246](https://github.com/pbek/QOwnNotes/issues/246))
 
 ## 16.07.7
 - fixed a possible crash after removing the last script in the settings dialog